backend/requirements.txt
# FastAPI
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.7.4,<3.0.0
python-multipart>=0.0.6

langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langchain-core>=0.3.0,<0.4.0

# Neo4j (with vector search support)
neo4j>=5.25.0,<6.0.0
langchain-neo4j>=0.1.0

# OpenAI
openai>=1.30.0

# Meter Reading (YOLO)
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0

# Utilities
numpy==1.24.3
python-dotenv>=1.0.0
